requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
jsonref==0.2
jsonschema==4.5.1
lark-parser==0.12.0
mappyfile==0.9.7
pyrsistent==0.18.1
requests==2.27.1
urllib3==1.26.9
